+++ web/src/containers/Nav.test.js
@@ -3,25 +3,9 @@ import React from 'react';
 import { BrowserRouter as Router } from 'react-router-dom'
 
 import {
-  NavLink,
   plain as Nav,
   mapStateToProps
 } from './Nav';
-
-describe('<NavLink /> component', () => {
-  it('renders props and children', () => {
-    const props = {
-      id: "executive-summary-summary-nav",
-      href: "/apd/executive-summary#executive-summary-summary",
-      otherThing: true
-    }
-    const component = shallow(<NavLink {...props}>Activities Summary</NavLink>)
-    expect(component.prop('id')).toEqual('executive-summary-summary-nav')
-    expect(component.prop('to')).toEqual('/apd/executive-summary#executive-summary-summary')
-    expect(component.prop('otherThing')).toEqual(true)
-    expect(component.text()).toEqual('Activities Summary')
-  });
-});
 
 describe('<Nav /> component', () => {
   it('renders correctly', () => {
